pipeline_components/requirements.txt
sklearn
pandas
seaborn
matplotlib
xgboost
google-api-python-client
dask[dataframe]
gcsfs
google-cloud
google-cloud-storage